refactor: deduplicate token exchange — single sync implementation shared by spawner and env listing
